componentize referral preferrences
